exceptions: introduce RustException
It's going to be used in the Rust wrapper. Whenever the Rust code
returns an error, we will stringify it and throw within this exception.

This is the first, simplest approach to handle errors from Rust. Later we can think of
more sophisticated ways to handle errors, e.g. by introducing different exception types
for different error cases.